requirements.txt
fastapi==0.65.1
uvicorn==0.14.0
pymongo==4.3.2
loguru==0.6.0
aioredis==2.0.1
motor==3.1.1
python-dotenv==0.21.0
asgiref==3.4.1
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
h11==0.12.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
starlette==0.14.2
typing-extensions==3.10.0.0